requirements.txt
Pillow
pytesseract
streamlit
os
pdf2image
poppler
json
